requirements.txt
django==4.1.*
djangorestframework==3.14.*
coverage==5.*
Pillow==9.2.*
requests==2.28.*